refactor: extract chat_turn_payload to turn-core
- chat_turn_payload.rs (336 LOC) extracted — all deps already in turn-core
- Runtime module replaced with re-export
