+++ rust/crates/astra-turn-core/src/chat_turn_payload.rs
@@ -1,0 +1,336 @@
+//! Top-level `/chat` (streaming) JSON body skeleton before `edge_tools`, `tool_results`, and selector hints.
+
+use std::collections::HashSet;
+use std::path::Path;
+
+use serde_json::{Value, json};
+
+use crate::chat_turn_edge_profile::build_base_edge_profile_value;
+use crate::chat_turn_explain_wire::chat_turn_explain_field_json;
+use crate::edge_prompt_context::detect_workspace_context;
+use crate::tool_schema_prune::filter_tool_schemas_by_excluded_names;
+
+/// Inputs for [`chat_turn_base_payload`] (keeps the arity aligned with the JSON body without a 9-arg function).
+pub struct ChatTurnBasePayloadInput<'a> {
+    pub messages: &'a [Value],
+    pub session_id: Option<&'a str>,
+    pub agent_id: Option<&'a str>,
+    pub model: Option<&'a str>,
+    pub explain_verbose: bool,
+    pub explain_on: bool,
+    pub edge_executor_id: &'a str,
+    pub capabilities: Vec<String>,
+    pub project_root: &'a Path,
+    pub git_branch: Option<String>,
+    /// Optional thinking/reasoning budget in tokens. When Some, the server should
+    /// constrain extended-thinking output to this many tokens.
+    pub thinking_budget_tokens: Option<u32>,
+}
+
+/// Invariant fields for `POST /chat/stream` (or equivalent) before dynamic tool schemas and callbacks.
+///
+/// `edge_executor_id` and `capabilities` are fields so server-side builders are not tied to CLI env.
+#[must_use]
+pub fn chat_turn_base_payload(input: ChatTurnBasePayloadInput<'_>) -> Value {
+    let ChatTurnBasePayloadInput {
+        messages,
+        session_id,
+        agent_id,
+        model,
+        explain_verbose,
+        explain_on,
+        edge_executor_id,
+        capabilities,
+        project_root,
+        git_branch,
+        thinking_budget_tokens,
+    } = input;
+    let mut payload = json!({
+        "messages": messages,
+        "session_id": session_id,
+        "agent_id": agent_id,
+        "model": model,
+        "explain": chat_turn_explain_field_json(explain_verbose, explain_on),
+        "edge_executor_id": edge_executor_id,
+        "capabilities": capabilities,
+        "edge_profile": build_base_edge_profile_value(
+            project_root.to_string_lossy().as_ref(),
+            git_branch,
+            detect_workspace_context(project_root),
+        ),
+    });
+    if let Some(budget) = thinking_budget_tokens {
+        if let Some(obj) = payload.as_object_mut() {
+            obj.insert("thinking_budget_tokens".to_string(), json!(budget));
+        }
+    }
+    payload
+}
+
+/// Set `edge_profile.active_skills` when the edge detected system skill hints in the user message.
+pub fn merge_active_skills_into_edge_profile(payload: &mut Value, active_skills: &[&str]) {
+    if active_skills.is_empty() {
+        return;
+    }
+    if let Some(root) = payload.as_object_mut()
+        && let Some(ep) = root.get_mut("edge_profile")
+        && let Some(ep_obj) = ep.as_object_mut()
+    {
+        ep_obj.insert("active_skills".to_string(), json!(active_skills));
+    }
+}
+
+/// Deduped skill names that affected this `/chat` request: selector-chosen registry skills,
+/// message-detected system skills ([`super::chat_turn_edge_profile::detect_active_system_skills_in_message`]),
+/// and registry skills whose instruction bodies were merged successfully.
+pub fn merge_invoked_skills_into_edge_profile(payload: &mut Value, invoked_skills: &[String]) {
+    if invoked_skills.is_empty() {
+        return;
+    }
+    if let Some(root) = payload.as_object_mut()
+        && let Some(ep) = root.get_mut("edge_profile")
+        && let Some(ep_obj) = ep.as_object_mut()
+    {
+        ep_obj.insert("invoked_skills".to_string(), json!(invoked_skills));
+    }
+}
+
+/// Shallow-merge top-level keys from `extensions` into `edge_profile` (cloud–edge audit / lineage).
+///
+/// `extensions` must be a JSON object. Non-object values are ignored.
+pub fn merge_edge_profile_extensions(payload: &mut Value, extensions: &Value) {
+    let Some(ext_obj) = extensions.as_object() else {
+        return;
+    };
+    if ext_obj.is_empty() {
+        return;
+    }
+    if let Some(root) = payload.as_object_mut()
+        && let Some(ep) = root.get_mut("edge_profile")
+        && let Some(ep_obj) = ep.as_object_mut()
+    {
+        for (k, v) in ext_obj {
+            ep_obj.insert(k.clone(), v.clone());
+        }
+    }
+}
+
+/// Dynamic tool schemas for this turn (`edge_tools`).
+pub fn set_payload_edge_tools(payload: &mut Value, schemas: Vec<Value>) {
+    if let Some(obj) = payload.as_object_mut() {
+        obj.insert("edge_tools".to_string(), Value::Array(schemas));
+    }
+}
+
+/// Drop schemas whose `function.name` is in `restricted_tools`, then set `edge_tools` on the payload.
+pub fn attach_filtered_edge_tools(
+    payload: &mut Value,
+    turn_schemas: Vec<Value>,
+    restricted_tools: &HashSet<String>,
+) {
+    let final_schemas = filter_tool_schemas_by_excluded_names(turn_schemas, restricted_tools);
+    set_payload_edge_tools(payload, final_schemas);
+}
+
+/// Callback-style `tool_results` array (only set when non-empty, matching historical CLI behavior).
+pub fn set_payload_tool_results_if_non_empty(payload: &mut Value, rows: &[Value]) {
+    if rows.is_empty() {
+        return;
+    }
+    if let Some(obj) = payload.as_object_mut() {
+        obj.insert("tool_results".to_string(), Value::Array(rows.to_vec()));
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::collections::HashSet;
+    use std::path::Path;
+
+    #[test]
+    fn base_payload_core_fields() {
+        let msgs = vec![json!({"role": "user", "content": "hi"})];
+        let p = chat_turn_base_payload(ChatTurnBasePayloadInput {
+            messages: &msgs,
+            session_id: None,
+            agent_id: Some("test-agent"),
+            model: Some("gpt-test"),
+            explain_verbose: false,
+            explain_on: true,
+            edge_executor_id: "edge-unit",
+            capabilities: vec!["bash".into(), "fs".into()],
+            project_root: Path::new("/tmp"),
+            git_branch: Some("main".into()),
+            thinking_budget_tokens: None,
+        });
+        assert_eq!(p["messages"], json!(msgs));
+        assert_eq!(p["session_id"], Value::Null);
+        assert_eq!(p["agent_id"], "test-agent");
+        assert_eq!(p["model"], "gpt-test");
+        assert_eq!(p["explain"], json!(true));
+        assert_eq!(p["edge_executor_id"], "edge-unit");
+        let caps = p["capabilities"].as_array().unwrap();
+        assert!(caps.contains(&json!("bash")));
+        assert_eq!(p["edge_profile"]["cwd"], "/tmp");
+        assert_eq!(p["edge_profile"]["git_branch"], "main");
+        assert!(p["edge_profile"].get("memoria_url").is_some());
+        // thinking_budget_tokens = None → field absent
+        assert!(p.get("thinking_budget_tokens").is_none());
+    }
+
+    #[test]
+    fn base_payload_session_id_some() {
+        let p = chat_turn_base_payload(ChatTurnBasePayloadInput {
+            messages: &[],
+            session_id: Some("sess-1"),
+            agent_id: None,
+            model: Some("m"),
+            explain_verbose: true,
+            explain_on: false,
+            edge_executor_id: "e",
+            capabilities: vec![],
+            project_root: Path::new("/"),
+            git_branch: None,
+            thinking_budget_tokens: None,
+        });
+        assert_eq!(p["session_id"], "sess-1");
+        assert_eq!(p["agent_id"], Value::Null);
+        assert_eq!(p["explain"], json!("verbose"));
+    }
+
+    #[test]
+    fn base_payload_thinking_budget_included_when_some() {
+        let p = chat_turn_base_payload(ChatTurnBasePayloadInput {
+            messages: &[],
+            session_id: None,
+            agent_id: None,
+            model: Some("claude-thinking"),
+            explain_verbose: false,
+            explain_on: false,
+            edge_executor_id: "e",
+            capabilities: vec![],
+            project_root: Path::new("/"),
+            git_branch: None,
+            thinking_budget_tokens: Some(10000),
+        });
+        assert_eq!(p["thinking_budget_tokens"], 10000);
+    }
+
+    #[test]
+    fn base_payload_thinking_budget_absent_when_none() {
+        let p = chat_turn_base_payload(ChatTurnBasePayloadInput {
+            messages: &[],
+            session_id: None,
+            agent_id: None,
+            model: Some("gpt-4o"),
+            explain_verbose: false,
+            explain_on: false,
+            edge_executor_id: "e",
+            capabilities: vec![],
+            project_root: Path::new("/"),
+            git_branch: None,
+            thinking_budget_tokens: None,
+        });
+        assert!(p.get("thinking_budget_tokens").is_none());
+    }
+
+    #[test]
+    fn merge_active_skills_into_edge_profile_inserts_array() {
+        let mut p = json!({ "edge_profile": {} });
+        merge_active_skills_into_edge_profile(&mut p, &["markdown", "concise"]);
+        assert_eq!(
+            p["edge_profile"]["active_skills"],
+            json!(["markdown", "concise"])
+        );
+    }
+
+    #[test]
+    fn merge_active_skills_no_op_when_empty() {
+        let mut p = json!({ "edge_profile": {} });
+        merge_active_skills_into_edge_profile(&mut p, &[]);
+        assert!(
+            p["edge_profile"]
+                .as_object()
+                .unwrap()
+                .get("active_skills")
+                .is_none()
+        );
+    }
+
+    #[test]
+    fn merge_invoked_skills_into_edge_profile_inserts_array() {
+        let mut p = json!({ "edge_profile": {} });
+        merge_invoked_skills_into_edge_profile(&mut p, &["markdown".into(), "bash".into()]);
+        assert_eq!(
+            p["edge_profile"]["invoked_skills"],
+            json!(["markdown", "bash"])
+        );
+    }
+
+    #[test]
+    fn merge_invoked_skills_no_op_when_empty() {
+        let mut p = json!({ "edge_profile": {} });
+        merge_invoked_skills_into_edge_profile(&mut p, &[]);
+        assert!(
+            p["edge_profile"]
+                .as_object()
+                .unwrap()
+                .get("invoked_skills")
+                .is_none()
+        );
+    }
+
+    #[test]
+    fn merge_edge_profile_extensions_merges_objects() {
+        let mut p = json!({ "edge_profile": { "cwd": "/tmp", "k": 1 } });
+        merge_edge_profile_extensions(
+            &mut p,
+            &json!({
+                "session_lineage": { "parent_session_id": "abc" },
+                "edge_policy": { "permission_mode": "prompt" }
+            }),
+        );
+        assert_eq!(p["edge_profile"]["cwd"], "/tmp");
+        assert_eq!(p["edge_profile"]["k"], 1);
+        assert_eq!(
+            p["edge_profile"]["session_lineage"]["parent_session_id"],
+            "abc"
+        );
+        assert_eq!(
+            p["edge_profile"]["edge_policy"]["permission_mode"],
+            "prompt"
+        );
+    }
+
+    #[test]
+    fn set_payload_edge_tools_and_tool_results() {
+        let mut p = json!({});
+        set_payload_edge_tools(&mut p, vec![json!({"fn": "t1"})]);
+        assert_eq!(p["edge_tools"], json!([{"fn": "t1"}]));
+        set_payload_tool_results_if_non_empty(&mut p, &[json!({"tool_call_id": "1"})]);
+        assert_eq!(p["tool_results"], json!([{"tool_call_id": "1"}]));
+    }
+
+    #[test]
+    fn set_payload_tool_results_no_op_when_empty() {
+        let mut p = json!({});
+        set_payload_tool_results_if_non_empty(&mut p, &[]);
+        assert!(p.get("tool_results").is_none());
+    }
+
+    #[test]
+    fn attach_filtered_edge_tools_excludes_by_name() {
+        let mut p = json!({});
+        let schemas = vec![
+            json!({"function": {"name": "bash"}}),
+            json!({"function": {"name": "danger"}}),
+        ];
+        let mut r = HashSet::new();
+        r.insert("danger".into());
+        attach_filtered_edge_tools(&mut p, schemas, &r);
+        let arr = p["edge_tools"].as_array().unwrap();
+        assert_eq!(arr.len(), 1);
+        assert_eq!(arr[0]["function"]["name"], "bash");
+    }
+}
